% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{avgrange}
\alias{avgrange}
\title{Average Across Range of a Tape}
\usage{
avgrange(pfun)
}
\arguments{
\item{pfun}{An \code{Rcpp_ADFun} object.}
}
\value{
An \code{Rcpp_ADFun} object.
}
\description{
Creates a \code{CppAD} tape that is the average of the returned values of \code{pfun}.
For creating this tape, the values of \code{pfun$dyntape} and \code{pfun$xtape} are used.
}
\seealso{
Other tape builders: 
\code{\link{fixdynamic}()},
\code{\link{fixindependent}()},
\code{\link{keeprange}()},
\code{\link{tape_Hessian}()},
\code{\link{tape_Jacobian}()},
\code{\link{tape_bdryw}()},
\code{\link{tape_gradoffset}()},
\code{\link{tape_logJacdet}()},
\code{\link{tape_smd}()},
\code{\link{tape_swap}()},
\code{\link{tape_uld}()}
}
\concept{tape builders}
