% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bingham.R
\name{Bingham}
\alias{Bingham}
\title{Score matching estimates for the Bingham distribution}
\usage{
Bingham(sample, A = NULL, method = "smfull", control = default_Rcgmin())
}
\arguments{
\item{sample}{A matrix of observations in Cartesian coordinates. Each row is a (multivariate) measurement.}

\item{A}{For full score matching only: if supplied, then NA elements of \code{A} are estimated and the other elements are fixed.}

\item{method}{The estimating method, either "smfull" for score matching estimates for all parameters
or "Mardia" for the \insertCite{mardia2016sc}{scorecompdir} hybrid estimator.}

\item{control}{Control parameters passed to \code{\link[Rcgmin:Rcgmin]{Rcgmin::Rcgmin()}}.}
}
\description{
Score matching estimator of the Bingham matrix.
}
\details{
The Bingham distribution , which has a density proportional to
\deqn{\exp(z^T A z),}
where \eqn{A} is a symmetric matrix and the trace (sum of the diagonals) of \eqn{A} is zero for identifiability \insertCite{@p181, @mardia2000di}{scorecompdir}.

The function \code{Bingham()} calculates the final diagonal element of \eqn{A} from the sum of the other diagonal elements to ensure that the trace of \eqn{A} is zero. All other elements of \eqn{A} are estimated using score matching.

The estimating method by \insertCite{mardia2016sc;textual}{scorecompdir} first calculates the maximum-likelihood estimate of the eigenvectors \eqn{G} of \eqn{A}.
The observations \code{Y} are then standardised to \code{Y}\eqn{G}.
After this standardisation, the non-diagonal elements of \eqn{A} are zero.
The diagonal elements are estimated using score matching.
See \insertCite{mardia2016sc}{scorecompdir} for details.
}
\examples{
p <- 4
A <- rsymmetricmatrix(p)
A[p,p] <- -sum(diag(A)[1:(p-1)]) #to satisfy the trace = 0 constraint
sample <- simdd::rBingham(100, A)

Bingham(sample, method = "Mardia")
}
\seealso{
Other directional model estimators: 
\code{\link{FB}()},
\code{\link{vMF_kappa}()},
\code{\link{vMF_m}()},
\code{\link{vMF_robust}()},
\code{\link{vMF}()}
}
\concept{directional model estimators}
