% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stiefel_utils.R
\name{commutation_mat}
\alias{commutation_mat}
\title{Commutation matrix}
\usage{
commutation_mat(...)
}
\arguments{
\item{m}{Number of rows of matrix \code{A}}

\item{p}{Number of columns of matrix \code{A}}
}
\value{
Matrix of size mp x mp.
}
\description{
The \insertCite{magnus2019ma}{scorematchingad} commutation matrix \eqn{K_{m,p}} that is such that
\code{K \%*\% vec(A) == vec(t(A))} for any m x p matrix \code{A}.
}
\details{
Uses memoisation to avoid recalculating the matrix for the same set of dimensions
}
