% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cppad_closed.R
\name{cppad_closed}
\alias{cppad_closed}
\title{Score Matching Estimator for Quadratic-Form Score-Matching Objective Functions}
\usage{
cppad_closed(
  smotape,
  Y,
  Yapproxcentres = NA * Y,
  w = rep(1, nrow(Y)),
  approxorder = 10
)
}
\arguments{
\item{smotape}{A \code{CppAD} tape of a score matching discrepancy function that has \emph{quadratic form}. Test for quadratic form using \code{\link[=testquadratic]{testquadratic()}}.
The \code{smotape}'s independent variables are assumed to be the model parameters to fit
and the \code{smotape}'s dynamic parameter is a (multivariate) measurement.}

\item{Y}{A matrix of multivariate observations. Each row is an observation.}

\item{Yapproxcentres}{A matrix of Taylor approximation centres for rows of Y that require approximation. \code{NA} for rows that do not require approximation.}

\item{w}{Weights for each observation.}

\item{approxorder}{The order of Taylor approximation to use.}
}
\description{
For a \code{CppAD} tape of a quadratic-form score matching discrepancy function, calculates the vector of parameters such that the gradient of the score matching discrepancy is zero.
Also estimates standard errors and covariance.
Many score matching estimators have an objective function that has a quadratic form.
}
\details{
When the score matching function is of quadratic form, then the gradient is zero at \eqn{H^{-1}b}{solve(H) \%*\% b},
where \eqn{H} is the average of the Hessian of the score matching function evaluated at each measurement and
\eqn{b} is the average of the gradient offset (see \code{\link[=quadratictape_parts]{quadratictape_parts()}}) evaluated at each measurement.
Both the Hessian and the gradient offset are constant with respect to the model parameters for quadratic-form score matching equations.

Standard errors are only computed when the weights are constant, and use the Godambe information matrix (aka sandwich method) (\emph{would like a good reference for this here}).
The sensitivity matrix \eqn{G} is estimated as
the negative of the average over the Hessian of \code{smotape} evaluated at each observation in \code{Y}.
The variability matrix \eqn{J} is then estimated as
the sample covariance (denominator of \eqn{n-1}) of the gradiant of \code{smotape} evaluated at each of the observations in \code{Y} for the estimated \eqn{\theta}.
The variance of the estimator is then estimated as
\eqn{G^{-1}JG^{-1}/n,}
where \code{n} is the number of observations.
}
\seealso{
Other generic score matching tools: 
\code{\link{Windham}()},
\code{\link{cppad_search}()}
}
\concept{generic score matching tools}
