% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numoptimisers.R
\name{default_Rcgmin}
\alias{default_Rcgmin}
\alias{default_FixedPoint}
\title{Default control parameters for numerical optimisers}
\usage{
default_Rcgmin()

default_FixedPoint()
}
\description{
This package uses \code{\link[Rcgmin:Rcgmin]{Rcgmin::Rcgmin()}} in many of the score matching estimates and
\code{default_Rcgmin()} returns the default control values used throughout the package.
Windham robustness (see \code{\link[=Windham]{Windham()}}) uses \code{\link[FixedPoint:FixedPoint]{FixedPoint::FixedPoint()}} for a fixed point search, with default control parameters given by \code{default_FixedPoint()}.
}
\details{
The default for \code{Rcgmin()} is \code{list(tol = 1E-20)}, which means the optimisation won't end until the squared size of the gradient summed over all observations at the estimate parameter set is less than 1E-20.
}
\seealso{
Other generic score matching tools: 
\code{\link{Windham}()},
\code{\link{cppad_closed}()},
\code{\link{cppad_search}()}
}
\concept{generic score matching tools}
