% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{evalll}
\alias{evalll}
\title{Evaluate a log-likelihood function}
\usage{
evalll(ll, u, theta)
}
\arguments{
\item{ll}{A compiled log-likelihood function created by \code{\link[=customll]{customll()}}.}

\item{u}{A vector of measurements for an individual}

\item{theta}{A vector of parameters}
}
\value{
The value of the log-likelihood at \code{u} with parameters \code{theta}.
}
\description{
Evaluate a log-likelihood function
}
