% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaltape.R
\name{evaltape}
\alias{evaltape}
\alias{evaltape_wsum}
\title{Simple wrapper to evaluate CppAD tapes and derivatives many times}
\usage{
evaltape(tape, xmat, pmat, xcentres = NA * xmat, approxorder = 10)

evaltape_wsum(
  tape,
  xmat,
  pmat,
  w = NULL,
  xcentres = NA * xmat,
  approxorder = 10
)
}
\arguments{
\item{tape}{An \code{\link{ADFun}} object (i.e. a tape of a function).}

\item{xmat}{A matrix of (multivariate) independent variables. Each represents a single independent variable vector. Or a single vector of that is used for all rows of \code{pmat}.}

\item{pmat}{A matrix of dynamic parameters. Or a single vector of dynamic parameters to use for all rows of \code{xmat}.}

\item{xcentres}{A matrix of approximation for Taylor approximation centres for \code{xmat}. Use values of \code{NA} for rows that do not require Taylor approximation.}

\item{approxorder}{Order of Taylor approximation}

\item{w}{Weights to apply to each row of \code{xmat} for computing the weighted sum. If \code{NULL} then each row is given a weight of \code{1}.}
}
\value{
A matrix, each row corresponding to the evaluation of the same row in \code{xmat}, \code{pmat} and the \code{xcentres}.
}
\description{
Evaluates a tape exactly or approximately for an array of provided variable values and dynamic parameter values.
The function \code{evaltape_wsum()} computes the column-wise weighted sum of the result.
}
\details{
Approximation is via Taylor approximation of the independent variable around the approximation centre provided in \code{xcentres}.
}
\seealso{
Other tape evaluators: 
\code{\link{quadratictape_parts}()},
\code{\link{smvalues_tape}()},
\code{\link{testquadratictape}()}
}
\concept{tape evaluators}
