% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifoldtransform.R
\name{manifoldtransform}
\alias{manifoldtransform}
\title{Build an object specifying the manifold-transform pair.}
\usage{
manifoldtransform(start, tran, man)
}
\arguments{
\item{start}{The starting manifold. Used for checking that \code{tran} and \code{man} match.}

\item{tran}{The name of a transformation. Available transformations are
\itemize{
\item `srqt'
\item `alr'
\item `clr'
\item \verb{none' (or }identity')
}}

\item{man}{The name of the manifold that \code{tran} maps \code{start} to. Available manifolds are:
\itemize{
\item `sph' unit sphere
\item `Hn111' hyperplane normal to 1, 1, 1, 1, ...
\item `sim' simplex
\item `Euc' Euclidean space
}}
}
\value{
A named list with:
\itemize{
\item \code{tran} A object of type \code{Rcpp_transform_ad} representing the transform
\item \code{man} A object of type \code{Rcpp_man_ad} representing the manifold
}
}
\description{
Generate an object used to specify a manifold and transformation to the manifold. This specific to each type of data.
}
\details{
Only some combinations of \code{start}, \code{tran} and \code{man} are available because \code{tran} must map between \code{start} and \code{man}.
These combinations of \code{start}-\code{tran}-\code{man} are available:
\itemize{
\item sim-sqrt-sph
\item sim-identity-sim
\item sim-alr-Euc
\item sim-clr-Hn111
\item sph-identity-sph
}
}
\examples{
manifoldtransform("sim", "alr", "Euc")
}
