% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microdata.R
\docType{data}
\name{microdata}
\alias{microdata}
\title{Microbiome Data from Article Supplement}
\format{
A \code{data.frame} with 94 rows and 31 columns.
}
\source{
To be confirmed.
}
\usage{
microdata
}
\description{
This file was created from the file \code{microbiome_data.xlsx} in the supplementary information of the log-ratio draft manuscript.
It contains two more measurements than the \code{microbiome_data.xlsx} file in the supplementary information of \insertCite{scealy2022sc}{scorecompdir},
which were deemed outliers by Scealy and Wood, and have \code{IndividualID} of \code{2079} and \code{2280}.
}
\details{
This data is currently private, pending permission from the original creators of the data.

The file was created \code{microbiome_data.xlsx} using the below code

\if{html}{\out{<div class="sourceCode">}}\preformatted{microdata <- readxl::read_excel("microbiome_data.xlsx")
microdata <- as.data.frame(microdata)
save(microdata, file = "./data/microdata.rda")
}\if{html}{\out{</div>}}
}
\references{
\insertAllCited{}
}
\keyword{datasets}
