% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pForward0}
\alias{pForward0}
\title{Evaluate a CppAD tape}
\usage{
pForward0(pfun, x, dynparam)
}
\arguments{
\item{pfun}{Rcpp::XPtr to an ADFun with dynamic parameters}

\item{x}{A vector in the domain of the taped function.}

\item{dynparam}{a vector of the dynamic parameters.}
}
\value{
The value of \code{pfun} evaluated at \code{x} with parameters \code{dynparam}.
}
\description{
Evaluate a CppAD tape
}
