% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pHessian}
\alias{pHessian}
\title{The Hessian of recorded function.}
\usage{
pHessian(pfun, value, dynparam)
}
\arguments{
\item{pfun}{An Rcpp::XPtr to an ADFun object (i.e. a tape of a function)}

\item{value}{A vector in the domain of the taped function.}

\item{dynparam}{a vector of the dynamic parameters. If \code{pfun} has no dynamic parameters then set \code{dynparam = vector(mode = "numeric")}.}
}
\value{
The Hessian of pfun as a vector.
}
\description{
The Hessian of recorded function.
}
