% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pParameter}
\alias{pParameter}
\title{Indicate Constant Components of Range}
\usage{
pParameter(pfun)
}
\arguments{
\item{pfun}{A CppAD tape.}
}
\value{
A vector logical values. \code{TRUE} indicates that element of the tape result is constant.
}
\description{
Use \code{CppAD}'s \code{Parameter()} function for \code{ADFun} objects to see if the returned values of a tape are constant with respect to the independent values.
}
\details{
The \code{CppAD} function \href{https://cppad.readthedocs.io/en/latest/fun_property.html#parameter}{\code{Parameter(i)}} returns \code{TRUE} when the \code{i}th component of the range does not depend on the independent value
(the \code{i}th component may still depend on the value of the dynamic parameters - see \url{https://cppad.readthedocs.io/en/latest/glossary.html#dynamic} ).
}
