% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppi_examplemodels.R,
%   R/ppi_examplemodels_gengamma.R
\name{ppi_egmodel}
\alias{ppi_egmodel}
\alias{ppi_eggengamma}
\title{Load and simulate example PPI models}
\usage{
ppi_egmodel(n, maxden = 4)

ppi_eggengamma(n, maxden = 0)
}
\arguments{
\item{n}{Number of samples to generate}

\item{maxden}{the log(C) maximum in simulating a PPI model. See \code{\link[=rppi]{rppi()}}.}
}
\value{
A list:
\itemize{
\item \code{sample} A matrix of the simulated samples (\code{n} rows)
\item \code{p} The number of components of the model
\item \code{theta} The PPI parameter vector
\item \code{AL} The \eqn{A_L} parameter matrix
\item \code{bL} The \eqn{b_L} parameter vector
\item \code{beta} The \eqn{\beta} parameter vector
}
}
\description{
The function \code{ppi_egmodel()} simulates the 3-component PPI model from \insertCite{@Section 2.3, @scealy2022sc}{scorecompdir}. The function \code{ppi_eggengamma()} simulates from the 5-component model of \insertCite{@Section A.11, @scealy2022sc}{scorecompdir}, which is a generalised gamma model (i.e. \eqn{b_L} = 0).
Model parameters and the simulated sample are returned.
}
\examples{
ppi_egmodel(1000)
ppi_eggengamma(100)
}
\references{
\insertAllCited{}
}
