% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppi_parammats.R
\name{ppi_parammats}
\alias{ppi_parammats}
\title{Convert a PPI Parameter Vector to AL, bL and beta}
\usage{
ppi_parammats(paramvec)
}
\arguments{
\item{paramvec}{A PPI parameter vector, typically created by \code{\link[=ppi_paramvec]{ppi_paramvec()}} or as an output of \code{\link[=ppi]{ppi()}}.}
}
\value{
A named list of \eqn{A_L}, \eqn{b_L}, and \eqn{\beta}.
}
\description{
The parameters for the PPI model are the symmetric matrix \eqn{A_L},
the vector \eqn{b_L}, and the vector \eqn{\beta}.
However computationally, it is necessary to convert these into a single vector, which is typically performed using \code{\link[=ppi_paramvec]{ppi_paramvec()}}.
The function \code{\link[=ppi_parammats]{ppi_parammats()}} is roughly the reverse operation:
any vector created by \code{\link[=ppi_paramvec]{ppi_paramvec()}} is converted into \eqn{A_L}, \eqn{b_L} and \eqn{\beta}.
}
\examples{
vec <- ppi_paramvec(AL = rsymmetricmatrix(2), beta = c(-0.8, -0.7, 0))
ppi_parammats(vec)
}
\seealso{
Other PPI model tools: 
\code{\link{dppi}()},
\code{\link{ppi_egmodel}()},
\code{\link{ppi_paramvec}()},
\code{\link{ppi_robust_alrgengamma}()},
\code{\link{ppi_robust}()},
\code{\link{ppi_smvalues}()},
\code{\link{ppi_toAstar}()},
\code{\link{ppi}()},
\code{\link{rppi}()}
}
\concept{PPI model tools}
