% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppi_paramvec.R
\name{ppi_paramvec}
\alias{ppi_paramvec}
\title{User friendly preparation of parameter vector for PPI model}
\usage{
ppi_paramvec(
  p = NULL,
  AL = NULL,
  bL = NULL,
  Astar = NULL,
  beta = NULL,
  betaL = NULL,
  betap = NULL
)
}
\arguments{
\item{p}{The number of components. If \code{NULL} then \code{p} will be inferred from other inputs.}

\item{AL}{Either \code{NULL}, a p-1 x p-1 symmetric matrix, a number, or "diag".
If NULL then all \eqn{A_L} elements will be set to NA, and \code{\link[=ppi]{ppi()}} will estimate them.
If a single number, then \eqn{A_L} will be fixed as a matrix of the given value.
If "diag" then the non-diagonal elements of \eqn{A_L} will be fixed to 0, and the diagonal will be \code{NA} (and estimated by \code{\link[=ppi]{ppi()}}).}

\item{bL}{Either \code{NULL}, a number, or a vector of length p-1.
If \code{NULL}, then all elements of \eqn{b_L} will be set to \code{NA}, and \code{\link[=ppi]{ppi()}} will estimate them.
If a number, then \eqn{b_L} will be fixed at the supplied value.}

\item{Astar}{Either \code{NULL} or a p x p matrix.
If non-null, then overrides \code{AL} and \code{bL}.
If a matrix, all elements must be non-\code{NA} and \code{Astar} will be converted to \code{AL} and \code{bL}.}

\item{beta}{Either \code{NULL}, a number, or a vector of length p.
If NULL then all elements of \eqn{\beta} will be set to \code{NA}, and \code{\link[=ppi]{ppi()}} will estimate them.
If a number then the \eqn{\beta} elements will be fixed at the given number.}

\item{betaL}{Either \code{NULL}, a number, or a vector of length p-1.
If \code{NULL} then the 1...(p-1)th \eqn{\beta} elements will be set to \code{NA}, and \code{\link[=ppi]{ppi()}} will estimate them.
If a number then the 1...(p-1)th \eqn{\beta} elements fixed at the given number.}

\item{betap}{Either \code{NULL} or a number.
If \code{NULL} then the \code{p}th element of \eqn{\beta} will be set to \code{NA}, and \code{\link[=ppi]{ppi()}} will estimate it.
If a number, then the pth element of \eqn{\beta} will be fixed at the given value.}
}
\value{
A vector starting with the diagonal elements of \eqn{A_L}, then the off-diagonal elements extracted by \code{\link[=upper.tri]{upper.tri()}} (which extracts elements of \eqn{A_L} along each row, left to right, then top to bottom), then \eqn{b_L}, then \eqn{\beta}.
}
\description{
The PPI model is parametrised using a symmetric covariance-like matrix \eqn{A_L}, a location-like vector \eqn{b_L} and a set of Dirichlet exponents \eqn{\beta}. For \code{p} components, \eqn{A_L} has \code{p-1} rows, \eqn{b_L} is a vector with \code{p-1} elements and \eqn{\beta} is a vector with \code{p} elements.
These are converted to a single vector using \code{ppi_paramvec()} that can then be used for fitting with \code{\link[=ppi]{ppi()}}, where NA-valued elements are fitted and all other elements are fixed.
}
\examples{
ppi_paramvec(AL = "diag", bL = 0, betap = -0.5, p = 3)
}
