% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppi_robust.R
\name{ppi_robust}
\alias{ppi_robust}
\title{Robustly Estimate Parameters of the PPI Distribution}
\usage{
ppi_robust(Y, cW, ...)
}
\arguments{
\item{Y}{A matrix of measurements. Each row is a measurement, each component is a dimension of the measurement.}

\item{cW}{A vector of robustness tuning constants. Easy to build using \code{\link[=ppi_cW]{ppi_cW()}} and \code{\link[=ppi_cW_auto]{ppi_cW_auto()}}. See \code{\link[=Windham]{Windham()}} for more details on \code{cW}.}

\item{...}{Passed to \code{\link[=Windham]{Windham()}} then to \code{\link[=ppi]{ppi()}}.}
}
\value{
A list:
\itemize{
\item \code{est} The estimated parameters as vector form (\code{paramvec}) and as \code{AL}, \code{bL} and \code{beta}.
\item \code{SE} "Not calculated." Returned for consistency with other estimators.
\item \code{info} Information returned in the \code{optim} slot of \code{\link[=Windham]{Windham()}}. Includes the final weights in \code{finalweights}.
}
}
\description{
Uses \code{\link[=Windham]{Windham()}} and \code{\link[=ppi]{ppi()}} to estimate a PPI distribution robustly.
There are many arguments to the \code{\link[=ppi]{ppi()}} function and we highly recommend trialling your arguments on \code{\link[=ppi]{ppi()}} first before running \code{ppi_robust()}.
}
\seealso{
Other PPI model tools: 
\code{\link{dppi}()},
\code{\link{ppi_egmodel}()},
\code{\link{ppi_parammats}()},
\code{\link{ppi_paramvec}()},
\code{\link{ppi_robust_alrgengamma}()},
\code{\link{ppi_smvalues}()},
\code{\link{ppi}()},
\code{\link{rppi}()}

Other Windham robustness functions: 
\code{\link{Windham}()},
\code{\link{ppi_robust_alrgengamma}()},
\code{\link{vMF_robust}()}
}
\concept{PPI model tools}
\concept{Windham robustness functions}
