% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppi_robust_alrgengamma.R
\name{ppi_robust_alrgengamma}
\alias{ppi_robust_alrgengamma}
\title{Windham Robustness for Scealy et al 2023}
\usage{
ppi_robust_alrgengamma(
  Y,
  cW,
  ...,
  fpcontrol = list(Method = "Simple", ConvergenceMetricThreshold = 1e-10)
)
}
\arguments{
\item{Y}{A matrix of measurements. Each row is a measurement, each component is a dimension of the measurement.}

\item{cW}{A vector of robustness tuning constants. When computing the weight for an observation the parameter vector is multiplied element-wise with \code{cW}. For the PPI model, generate \code{cW} easily using \code{\link[=ppi_cW]{ppi_cW()}} and \code{\link[=ppi_cW_auto]{ppi_cW_auto()}}.}

\item{...}{Passed to \code{\link[=Windham]{Windham()}} and on to \code{\link[=ppi]{ppi()}}.}

\item{fpcontrol}{A named list of control arguments to pass to \code{\link[FixedPoint:FixedPoint]{FixedPoint::FixedPoint()}} for the fixed point iteration.}
}
\value{
A list:
\itemize{
\item \code{est} The estimated parameters as vector form (\code{paramvec}) and as \code{AL}, \code{bL} and \code{beta}.
\item \code{SE} "Not calculated." Returned for consistency with other estimators.
\item \code{info} Information returned in the \code{optim} slot of \code{\link[=Windham]{Windham()}}. Includes the final weights in \code{finalweights}.
}
}
\description{
Performs the Windham robustification algorithm exactly as described in \insertCite{scealy2024ro;textual}{scorecompdir} for score matching via log-ratio transform of the PPI model with \eqn{b_L = 0}. This function calls the more general \code{\link[=Windham]{Windham()}} and \code{\link[=ppi]{ppi()}}.
}
\details{
This method must fit a PPI model via additive-log ratio transform with \eqn{b_L=0} fixed and the final element of \eqn{\beta} fixed.
The default convergence metric and threshold are different to the default for \code{\link[=ppi_robust]{ppi_robust()}} to match the original implementation: convergence is measured by the change in the first element of \eqn{\beta}, and convergence is reached when the change is smaller than \code{1E-6}. Override this behaviour by specifying the elements \code{ConvergenceMetric} and \code{ConvergenceMetricThreshold} in a list passed as \code{fpcontrol}.
\code{\link[=Windham]{Windham()}} is called with \code{alternative_populationinverse = TRUE}.
}
\references{
\insertAllCited{}
}
\seealso{
Other PPI model tools: 
\code{\link{dppi}()},
\code{\link{ppi_parammats}()},
\code{\link{ppi_paramvec}()},
\code{\link{ppi_robust}()},
\code{\link{ppi_smvalues}()},
\code{\link{ppi_toAstar}()},
\code{\link{ppi}()},
\code{\link{rppi}()}

Other Windham robustness functions: 
\code{\link{Windham}()},
\code{\link{ppi_robust}()},
\code{\link{vMF_robust}()}
}
\concept{PPI model tools}
\concept{Windham robustness functions}
