% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppi_toAStar.R
\name{ppi_toAstar}
\alias{ppi_toAstar}
\alias{ppi_fromAstar}
\title{Obtain AL and bL from Astar and Reverse}
\usage{
ppi_toAstar(AL, bL)

ppi_fromAstar(Astar)
}
\arguments{
\item{AL}{The \code{p-1} by \code{p-1} \eqn{A_L} matrix}

\item{bL}{The vector \eqn{b_L} (length \code{p-1}).}

\item{Astar}{The \eqn{A^*} matrix (a p by p symmetric matrix)}
}
\description{
\insertCite{scealy2022sc;textual}{scorecompdir} give two parameterisations of the PPI model. The function \code{\link[=ppi]{ppi()}} and most other PPI functions in this package use the parameterisation with matrix \eqn{A_L} and vector \eqn{b_L}. An alternative parameterisation uses a single matrix \code{Astar} instead of \eqn{A_L} and \eqn{b_L}.
}
\details{
The \code{Astar} parametrisation rewrites the PPI density as proportional to
\deqn{\exp(u^TA^*u)\prod_{i=1}^p u_i^{\beta_i},}
where \eqn{A^*} (\code{Astar}) is a \eqn{p} by \eqn{p} matrix.
Because \eqn{u} lies in the simplex (in particular \eqn{\sum u_i = 1}), the density is the same regardless of the value of \eqn{1^T A^* 1}=\code{sum(Astar)}, where \eqn{1} is the vector of ones. Thus \eqn{A_L} and \eqn{b_L} specify \eqn{A^*} up to an additive factor. In the conversion \code{toAstar()}, \eqn{A^*} is returned such that \eqn{1^T A^* 1 = 0}.
}
\section{Functions}{
\itemize{
\item \code{ppi_toAstar()}: Takes matrix \code{AL} and vector \code{bL} and returns the equivalent \code{Astar} with \eqn{1^T A^* 1 = }\code{sum(Astar) = 0}.

\item \code{ppi_fromAstar()}: Takes matrix \code{Astar} and returns the correponding \code{AL} and \code{bL} that give the same PPI density. It also returns the constant remainder that can be ignored due to normalisation of the probability density, but is useful for checking.

}}
\examples{
 Astar <- rWishart(1, 6, diag(3))[,,1]
 ppi_fromAstar(Astar)
 ppi_toAstar(ppi_fromAstar(Astar)$AL, ppi_fromAstar(Astar)$bL)
}
\seealso{
Other PPI model tools: 
\code{\link{dppi}()},
\code{\link{ppi_egmodel}()},
\code{\link{ppi_parammats}()},
\code{\link{ppi_paramvec}()},
\code{\link{ppi_robust_alrgengamma}()},
\code{\link{ppi_robust}()},
\code{\link{ppi_smvalues}()},
\code{\link{ppi}()},
\code{\link{rppi}()}
}
\concept{PPI model tools}
