% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcpp_ADFun_print.R
\name{print,Rcpp_ADFun}
\alias{print,Rcpp_ADFun}
\alias{print,Rcpp_ADFun-method}
\alias{show,Rcpp_ADFun-method}
\title{Print or show a summary of an Rcpp_ADFun}
\usage{
\S4method{print}{Rcpp_ADFun}(x, ...)

\S4method{show}{Rcpp_ADFun}(object)
}
\arguments{
\item{x}{An object of class \linkS4class{Rcpp_ADFun}.}

\item{...}{Passed to \code{\link[=format]{format()}}.}

\item{object}{An object of class \linkS4class{Rcpp_ADFun}.}
}
\description{
Both \code{print()} and \code{show()} will display a summary of a \linkS4class{Rcpp_ADFun} object.
}
\details{
The \code{show()} method overrides the default \code{show()} method for \code{\link[Rcpp:CppObject-class]{Rcpp::C++Object}} objects from the \code{Rcpp} package.
}
