% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pTapeGradOffset}
\alias{pTapeGradOffset}
\title{Tape the Gradient Offset of a Quadratic CppAD Tape}
\usage{
pTapeGradOffset(pfun, x, dynparam)
}
\arguments{
\item{pfun}{A quadratic CppAD Tape. Test for quadratic form using \code{\link[=testquadratictape]{testquadratictape()}}.}

\item{x}{A vector in the domain of the taped function.}

\item{dynparam}{a vector of the dynamic parameters}
}
\value{
A \code{Rcpp::XPtr} to a CppAD::ADFun object. The independent argument to the function are the dynamic parameters of \code{pfun}.
}
\description{
A quadratic function can be written as
\deqn{f(x;\theta) = \frac{1}{2} x^T W(\theta) x + b(\theta)^Tx + c.}
The function \code{pTapeGradOffset} creates a tape of \eqn{b(\theta)} where \eqn{\theta} is the independent variable.
}
\details{
The gradient of \eqn{f(x; \theta)} with respect to \eqn{\theta} is
\deqn{\Delta f(x; \theta) = \frac{1}{2}(W(\theta) + W(\theta)^T)\theta + b(\theta),}
and the Hessian is
\deqn{H f(x; \theta) = \frac{1}{2}(W(\theta) + W(\theta)^T),}
which does not depend on \eqn{x}.
The gradient of the function can be rewritten as
\deqn{\Delta f(x;\theta) = H f(x; \theta) x + b(\theta)^T x.}
The tape calculates \eqn{b(\theta)} as
\deqn{b(\theta) = \Delta f(x;\theta) - H f(x; \theta) x},
which does not depend on \eqn{x}.
In \code{pTapeGradOffset()} the \code{x} provided as an argument is used as the template for calculating \eqn{b(\theta)}.
The \code{x} vector and \code{dynparam} are used as the values to conduct the taping.
}
