% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ptapelogdetJ}
\alias{ptapelogdetJ}
\title{Tape the log of Jacobian determinant of a CppAD Tape}
\usage{
ptapelogdetJ(pfun, x, dynparam)
}
\arguments{
\item{pfun}{Rcpp::XPtr to an ADFun tape a tape with dynamic parameters and independent parameters}

\item{x}{A vector in the domain of the taped function.}

\item{dynparam}{a vector of the dynamic parameters}
}
\value{
A \code{Rcpp::XPtr} to a CppAD::ADFun object.
}
\description{
Creates a tape of the log of the Jacobian determinant of a function taped by CppAD.
The \code{x} vector is used as the value to conduct the taping.
}
