% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pTaylorApprox}
\alias{pTaylorApprox}
\title{The value of a recorded function approximated by Taylor expansion}
\usage{
pTaylorApprox(pfun, u, centre, dynparam, order)
}
\arguments{
\item{pfun}{Rcpp::XPtr to an ADFun tape a tape with independent values that are the points to be differentiated with}

\item{u}{A vector in the domain of the taped function.}

\item{centre}{A vector in the domain of the taped function to approximate the value at \code{u} from.}

\item{dynparam}{a vector of the dynamic parameters}

\item{order}{The order of Taylor expansion to use.}
}
\value{
The approximate value of pfun
}
\description{
Approximates the value of a \code{CppAD} tape at \code{u} using a Taylor approximation at \code{centre}. The dynamic parameters of the tape are set by \code{dynparam}.
}
