% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadratictape_parts.R
\name{quadratictape_parts}
\alias{quadratictape_parts}
\title{Evaluate the Hessian and gradient offset of a Taped Quadratic Function}
\usage{
quadratictape_parts(tape, tmat, tcentres = NA * tmat, approxorder = 10)
}
\arguments{
\item{tape}{A tape of a quadratic function (such as score matching objective function)}

\item{tmat}{A matrix of \code{t} vectors. Each row corresponds to a vector.}

\item{tcentres}{A matrix of Taylor approximation centres for rows of \code{tmat} that require approximation. \code{NA} for rows that do not require approximation.}

\item{approxorder}{The order of the Taylor approximation to use.}
}
\value{
A list of
\itemize{
\item \code{offset} Array of offsets \eqn{b(t)}, each row corresponding to a row in \code{tmat}
\item \code{Hessian} Array of offsets \eqn{b(t)}, each row corresponding to a row in \code{tmat}
}
}
\description{
When the score matching objective function is quadratic then the gradient of the score matching objective function can be written using the Hessian and an offset term. This can be useful for solving for the situation when the gradient is zero.
The Hessian and offset term are computed using \code{CppAD} tapes.
Taylor approximation can be used for locations at removable singularities.
}
\details{
A quadratic function can be written
\deqn{f(x; t) = \frac{1}{2} x^T W(t) x + b(t)^T x + c,}
where \eqn{t} is considered a vector that is constant with respect to the differentiation.
The Hessian of the function is with respect to \eqn{x} is
\deqn{H f(x; t) = \frac{1}{2}(W(t) + W(t)^T).}
The gradient of the function with respect to \eqn{x} can then be written
\deqn{\Delta f(x;t) = H f(x; t) x + b(t)^T x,}
where the Hessian and offset \eqn{b(t)} depend only on \eqn{t}.

The functions here evaluate the Hessian and offset \eqn{b(t)} for many values of \eqn{t}.
Tapes of the Hessian and gradient offset are created using \code{\link[=pTapeHessian]{pTapeHessian()}} and \code{\link[=pTapeGradOffset]{pTapeGradOffset()}} respectively.
These tapes are then evaluated for every row of \code{tmat}.
When the \code{tcentres} row is not \code{NA}, then approximate results are calculated using \code{\link[=pTaylorApprox]{pTaylorApprox()}}
}
