% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADFun.R
\docType{class}
\name{Rcpp_ADFun-class}
\alias{Rcpp_ADFun-class}
\alias{Rcpp_ADFun}
\title{A Class for CppAD Tapes}
\usage{
$eval(x, dyn)

$Jac(x, dyn)

$Hes(x, dyn)

$forward(q, x)

$Jacobian(x)

$Hessiani(x, i)

$Hessian0(x)

$Hessianw(x, w)

$new_dynamic(dyn)

$parameter(i)

$set_check_for_nan(bool)

$get_check_for_nan()

$domain

$range

$size_dyn_ind

$name

$xtape

$dyntape

$size_order
}
\arguments{
\item{x}{A vector of independent variables.}

\item{dyn}{A vector of dynamic parameters.}

\item{q}{Taylor coefficient order for evaluating derivatives with \verb{$forward()}.}

\item{i}{Index of range result.}

\item{name}{An easy to read name for the tape.}

\item{bool}{\code{TRUE} or \code{FALSE} to set \code{check_for_nan} behaviour using \verb{$set_check_for_nan()}.}

\item{w}{Weights assigned to each element of the range, for use with \verb{$Hessianw()}.}
}
\description{
Tapes are a record of operations performed by a function. Tapes can be evaluated and differentiated, and have properties (such as domain and range dimensions). Tapes also have dynamic parameters that can be updated. This class, \code{Rcpp_ADFun} uses `reference' semantics, so that copies all point to the same object and changes modify in place (i.e. changes modify the same object).
}
\details{
An objects of class \code{Rcpp_ADFun} wraps an \code{ADFun} object from \code{CppAD}. Many of the properties and behaviour of an \code{Rcpp_ADFun} object come directly from \code{ADFun} objects so more details and context can be found by looking at the \code{ADFun} object help in the \code{CppAD} \href{https://cppad.readthedocs.io}{\code{help}}.

Default printing of an \code{Rcpp_ADFun} object gives a short summary of the object and is implemented in the internal \code{print.Rcpp_ADFun()} function.
Technically the class name is 'Rcpp_ADFun' (so \code{inherits(x, "Rcpp_ADFun")} will return \code{TRUE}) and it is a reference class that connects to \code{CppAD} tapes in \verb{C++}. Many of the methods available for tapes in \code{CppAD} are made available here.

Tapes cannot be saved from session to session.
}
\section{Properties:}{
\itemize{
\item \code{size_order} Number of Taylor coefficient orders, per variable and direction, currently calculated and stored.
\item \code{domain} Dimension of the domain space (i.e., length of the independent variables vector).
\item \code{range} Dimension of the range space (i.e., length of the vector returned by \verb{$eval()}).
\item \code{size_dyn_ind} Number of independent dynamic parameters (i.e., length of the vector of dynamic parameters).
\item \code{name} An optional name for the tape.
\item \code{xtape} The values of the independent variables used for the initial taping.
\item \code{dyntape} The values of the dynamic parameters used for the initial taping.
\item \code{get_check_for_nan()} Debugging: Return whether the tape is configured to check for NaN values during computation. The check for NaN values only occurs if the \verb{C++} compilation enables debugging.
\item \code{set_check_for_nan(bool)} Set whether the tape should check for NaN values during computation (only effective if C++ debugging is enabled).
\item \code{parameter(i)} Check if the \code{i}th component of the range corresponds to a constant parameter.
}
}

\section{Methods:}{
\itemize{
\item \code{new_dynamic(dyn)} Specify new values for the dynamic parameters.
\item \code{forward(q, x)} Perform forward mode evaluation for the specified Taylor coefficient order \code{q}.
\item \code{Jacobian(x)} Evaluate the Jacobian of the function at the current set of dynamic parameters.
\item \code{Hessiani(x, i)} Evaluate the Hessian for the \code{i}-th element of the range (where \code{i = 0, 1, ...}).
\item \code{Hessian0()} Evaluate the Hessian for the first element of the range.
\item \code{Hessianw(x, w)} Evaluate the Hessian for a weighted sum of the range.
\item \code{eval(x, dyn)} Evaluate the function at new values of the variables and dynamic parameters.
\item \code{Jac(x, dyn)} Compute the Jacobian at new values of the variables and dynamic parameters.
\item \code{Hes(x, dyn)} Compute the Hessian (of the first element of the range) at new values of the variables and dynamic parameters.
}
}

\section{Extends}{
Extends class \linkS4class{C++Object} from the \code{Rcpp} package (\code{\link[Rcpp:CppObject-class]{Rcpp::C++Object}}), which is a \verb{reference class}.

For those familiar with \verb{C++}: An object of class \code{Rcpp_ADFun} contains a pointer to a \code{CppAD} \code{ADFun} object.
}

\section{Introduction to CppAD Tapes}{
This package uses version 2024000.5 of the algorithmic differentiation library \code{CppAD} \insertCite{bell2023cp}{scorematchingad} to build score matching estimators.
Full help for \code{CppAD} can be found at \url{https://cppad.readthedocs.io/}.

Differentiation proceeds by \emph{taping} the basic (\emph{atomic}) operations performed on the independent variables and dynamic parameters. The atomic operations include multiplication, division, addition, sine, cosine, exponential and many more.
Example values for the variables and parameters are used to conduct this taping, so care must be taken with any conditional (e.g. if-then) operations, and \href{https://cppad.readthedocs.io/}{\code{CppAD}} has a special tool for this called \code{CondExp} (short for \verb{conditional expressions}).

The result of taping is an object of class \code{ADFun} in \code{CppAD} and is often called a \emph{tape}.
This \code{ADFun} object can be evaluated, differentiated, used for further taping (via \code{CppAD}'s \code{base2ad()}), solving differential equations and more.
The differentiation is with respect to the independent variables, however the dynamic parameters can be altered which allows for creating a new \code{ADFun} object where the dynamic parameters become independent variables (see \code{\link[=tapeSwap]{tapeSwap()}}).
For the purposes of score matching, there are also \emph{fixed} parameters, which are the elements of the model's parameter vector that are given and not estimated.
}

\section{Warning: multiple CPU}{
Each time a tape is evaluated the corresponding \verb{C++} object is altered. Parallel use of the same \code{ADFun} object thus requires care and is not tested. For now I recommend creating a new \code{ADFun} object for each CPU.
}

\section{Improvements}{
A few methods for \code{CppAD} \code{ADFun} objects are not yet available through \code{Rcpp_ADFun} objects. These ones would be nice to include:
\itemize{
\item \code{optimize()}
}
}

\examples{
tape <- tape_uld_inbuilt("dirichlet", c(0.1, 0.4, 0.5), c(-0.5, -0.4, -0.2))
# Properties
tape$domain
tape$range
tape$size_dyn_ind
tape$name
tape$xtape
tape$dyntape
tape$size_order

# Convenient evaluation
tape$eval(x = c(0.2, 0.3, 0.5), dyn = c(-0.1, -0.1, -0.5))
tape$Jac(x = c(0.2, 0.3, 0.5), dyn = c(-0.1, -0.1, -0.5))
matrix(tape$Hes(x = c(0.2, 0.3, 0.5), dyn = c(-0.1, -0.1, -0.5)), nrow = tape$domain)

# Further methods
tape$forward(order = 1, x = c(0.2, 0.3, 0.5))
tape$Jacobian(x = c(0.2, 0.3, 0.5))
tape$Hessiani(x = c(0.2, 0.3, 0.5), i = 0)
tape$Hessian0(x = c(0.2, 0.3, 0.5))
tape$Hessianw(x = c(0.2, 0.3, 0.5), w = c(2))
tape$new_dynamic(dyn = c(-0.1, -0.1, -0.5))
tape$parameter(0)
tape$set_check_for_nan(FALSE)
tape$get_check_for_nan()
}
