% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADFun_print.R
\name{Rcpp_ADFun-methods}
\alias{Rcpp_ADFun-methods}
\alias{print,Rcpp_ADFun-method}
\alias{show,Rcpp_ADFun-method}
\title{Methods for Rcpp_ADFun Objects}
\usage{
\S4method{print}{Rcpp_ADFun}(x, ...)

\S4method{show}{Rcpp_ADFun}(object)
}
\arguments{
\item{x}{An object of class \linkS4class{Rcpp_ADFun}.}

\item{...}{Passed to \code{\link[=format]{format()}}.}

\item{object}{An object of class \linkS4class{Rcpp_ADFun}.}
}
\description{
Displays a summary of a \linkS4class{Rcpp_ADFun} object.

Displays a summary of a \linkS4class{Rcpp_ADFun} object and overrides the default \code{show()} method for \linkS4class{C++Object} objects from \code{\link[Rcpp:CppObject-class]{Rcpp::C++Object}}.
}
