% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{reembed}
\alias{reembed}
\title{Build unnormalised log-density in new embedding}
\usage{
reembed(uld, tran)
}
\arguments{
\item{tran}{A transform object.}

\item{pfun}{An \code{Rcpp_ADFun} object.}
}
\description{
Build a tape of the unnormalised log-density on a different embedding of the manifold.
}
\details{
When the embedding of a manifold changes according to \code{tran}, then the Riemannian metric on the manifold changes and so the uniform measure of the manifold also changes.
This change is accounted for using the \code{logdetJfromM()} property of transform objects.

The returned tape has an independent variable that is on the newly embedded manifold, and the value used for taping was \code{tran$toM(pfun$xtape)}.
}
