% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rppi.R
\name{rppi}
\alias{rppi}
\title{Generate random observations from the PPI model}
\usage{
rppi(n, ..., paramvec = NULL, maxden = 4, maxmemorysize = 1e+05)
}
\arguments{
\item{n}{Sample size}

\item{...}{
  Arguments passed on to \code{\link[=ppi_paramvec]{ppi_paramvec}}
  \describe{
    \item{\code{AL}}{Either \code{NULL}, a p-1 x p-1 symmetric matrix, a number, or "diag".
If NULL then all \eqn{A_L} elements will be set to NA, and \code{\link[=ppi]{ppi()}} will estimate them.
If a single number, then \eqn{A_L} will be fixed as a matrix of the given value.
If "diag" then the non-diagonal elements of \eqn{A_L} will be fixed to 0, and the diagonal will be \code{NA} (and estimated by \code{\link[=ppi]{ppi()}}).}
    \item{\code{bL}}{Either \code{NULL}, a number, or a vector of length p-1.
If \code{NULL}, then all elements of \eqn{b_L} will be set to \code{NA}, and \code{\link[=ppi]{ppi()}} will estimate them.
If a number, then \eqn{b_L} will be fixed at the supplied value.}
    \item{\code{Astar}}{Either \code{NULL} or a p x p matrix.
If non-null, then overrides \code{AL} and \code{bL}.
If a matrix, all elements must be non-\code{NA} and \code{Astar} will be converted to \code{AL} and \code{bL} using \code{\link[=ppi_fromAstar]{ppi_fromAstar()}}.}
    \item{\code{beta}}{Either \code{NULL}, a number, or a vector of length p.
If NULL then all elements of \eqn{\beta} will be set to \code{NA}, and \code{\link[=ppi]{ppi()}} will estimate them.
If a number then the \eqn{\beta} elements will be fixed at the given number.}
    \item{\code{betaL}}{Either \code{NULL}, a number, or a vector of length p-1.
If \code{NULL} then the 1...(p-1)th \eqn{\beta} elements will be set to \code{NA}, and \code{\link[=ppi]{ppi()}} will estimate them.
If a number then the 1...(p-1)th \eqn{\beta} elements fixed at the given number.}
    \item{\code{betap}}{Either \code{NULL} or a number.
If \code{NULL} then the \code{p}th element of \eqn{\beta} will be set to \code{NA}, and \code{\link[=ppi]{ppi()}} will estimate it.
If a number, then the pth element of \eqn{\beta} will be fixed at the given value.}
    \item{\code{p}}{The number of components. If \code{NULL} then \code{p} will be inferred from other inputs.}
  }}

\item{paramvec}{The PPI parameter vector, created easily using \code{\link[=ppi_paramvec]{ppi_paramvec()}} and also returned by \code{\link[=ppi]{ppi()}}.}

\item{maxden}{This is the constant \eqn{log(C)} in \insertCite{@Appendix A.1.3 @scealy2022sc}{scorecompdir}.}

\item{maxmemorysize}{Advanced use. The maximum size, in bytes, for matrices containing simulated Dirchlet samples. The default of \code{1E5} corresponds to 100 mega bytes.}
}
\value{
A matrix with \code{n} rows and \code{p} columns. Each row is an independent draw from the specified PPI distribution.
}
\description{
Given parameters of the PPI model, generates independent samples.
}
\details{
We recommend running \code{rppi()} a number of times to ensure the choice of \code{maxden} is good. \code{rppi()} will error when \code{maxden} is too low.

The simulation uses a rejection-sampling algorithm with Dirichlet proposal \insertCite{@Appendix A.1.3 @scealy2022sc}{scorecompdir}.
Initially \code{n} Dirichlet proposals are generated. After rejection there are fewer samples remaining, say \eqn{n^*}{n*}.
The ratio \eqn{n^*/n}{n*/n} is used to guess the number of new Dirichlet proposals to generate until \code{n} samples of the PPI model are reached.

Advanced use: The number of Dirichlet proposals created at a time is limited such that the matrices storing the Dirchlet proposals are always smaller than \code{maxmemorysize} bytes (give or take a few bytes for wrapping), which reduces the use of virtual RAM.
Larger \code{maxmemorysize} leads to faster simulation so long as \code{maxmemorysize} bytes are reliably contiguously available in RAM.
}
\examples{
beta0=c(-0.8, -0.8, -0.5)
AL = diag(nrow = 2)
bL = c(2, 3)
samp <- rppi(100,beta=beta0,AL=AL,bL=bL)
}
\seealso{
Other PPI model tools: 
\code{\link{dppi}()},
\code{\link{ppi_egmodel}()},
\code{\link{ppi_parammats}()},
\code{\link{ppi_paramvec}()},
\code{\link{ppi_robust_alrgengamma}()},
\code{\link{ppi_robust}()},
\code{\link{ppi_smvalues}()},
\code{\link{ppi_toAstar}()},
\code{\link{ppi}()}
}
\concept{PPI model tools}
