% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rRivest.R
\name{rRivest}
\alias{rRivest}
\title{Simulate from the Rivest model}
\usage{
rRivest(n, k, A, i)
}
\arguments{
\item{n}{Number of samples}

\item{k}{The concentration parameter for the Fisher component. Due to the sign convention for the eigenvector, this concentration parameter can be negative.}

\item{A}{A symmetric matrix with trace of zero.}

\item{i}{The eigenvector for the Fisher component is selected as the \code{i}th vector in increase eigenvalue magnitude.
E.g. \code{i=1} means the eigenvector with smallest eigenvalue size.}
}
\description{
This function uses \code{simdd::rFisherBingham()} to simulate from the Rivest model.
The direction of the Fisher compenent is chosen based on the magnitude of the eigenvalues using an index.
}
\details{
The Fisher component eigen vector is converted to have first element negative. Only the sphere is supported.
}
\examples{
p <- 3
A <- rsymmetricmatrix(p, -10, 10)
A[p,p] <- -sum(diag(A)[1:(p-1)]) #to satisfy the trace = 0 constraint for Bingham
rRivest(100, -3.2, A, 2)
}
