% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorecompdir-package.R
\docType{package}
\name{scorecompdir-package}
\alias{scorecompdir}
\alias{scorecompdir-package}
\title{scorecompdir: DRAFT Score-Matching for Directional and Compositional Data}
\description{
Score matching is a useful estimation technique when the normalising constant for a probability distribution is difficult to compute. This package includes methods for the flexible Polynomially-Tilted Pairwise Interaction (PPI) model for compositional data. These methods can perform well when there are zeros, even many zeros. See Scealy and Wood (2022) \url{https://doi.org/10.1080/01621459.2021.2016422} and Scealy, Hingee, Kent, and Wood (2024 - in publication). The package also includes hybrid estimators for directional distributions proposed by Mardia, Kent and Laha (2016) \url{http://arxiv.org/abs/1604.08470}. Tools for implementing additional score-matching estimators using CppAD \url{https://github.com/coin-or/CppAD} for algorithmic differentiation are also included.
}
\details{
This package's major contributions are
\itemize{
\item A general capacity to implement score matching estimators that use algorithmic differentiation to avoid tedious manual algebra.
The package uses \code{CppAD} and \code{Eigen} to differentiate model densities and compute the score matching objective (the \eqn{\hat\psi(f, f_0)} in the Introduction to Score Matching vignette \href{../doc/scorematchingintro.pdf}{\code{vignette("scorematchingintro")}}.
The score matching objective is minimised by a modern implementation of conjugate gradient descent available through \code{\link[optimx:Rcgmin]{optimx::Rcgmin()}}.
\item Score matching estimators for the Polynomially-Tilted Pairwise Interaction (PPI) model \insertCite{scealy2022sc}{scorecompdir}. See function \code{\link[=ppi]{ppi()}}.
\item Score matching and hybrid score matching estimators for von Mises Fisher and Bingham directional distributions \insertCite{mardia2016sc}{scorecompdir}.
\item Implementation of a modification of Windham's robustifying method \insertCite{windham1995ro}{scorecompdir} for many exponential family distributions. See \code{\link[=Windham]{Windham()}}.
For some models the density approaches infinity at some locations, creating difficulties for the weights in Windham's original method \insertCite{@scealy2022ro}{scorecompdir}.
\insertNoCite{*}{scorecompdir}
}
}
\references{
\insertAllCited{}
}
\author{
\strong{Maintainer}: Kassel Liam Hingee \email{kassel.hingee@anu.edu.au} (\href{https://orcid.org/0000-0001-9894-2407}{ORCID})

Authors:
\itemize{
  \item Janice Scealy
}

Other contributors:
\itemize{
  \item Bradley M. Bell [copyright holder]
  \item Ngunnawal Country [contributor]
  \item Ngambri Country [contributor]
}

}
