% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplex_bdryapprox.R
\name{simplex_boundaryshift}
\alias{simplex_boundaryshift}
\title{Generate approximation centres for measurements on the boundary of the simplex}
\usage{
simplex_boundaryshift(Y, shiftsize = 1e-04)
}
\arguments{
\item{Y}{The compositional measurements}

\item{shiftsize}{The distance to away from \code{u} to create the approximation centre.}
}
\value{
A matrix the same size as \code{Y}, containing the measurements in \code{Y} translated \code{shiftsize} distance towards the simplex centre.
}
\description{
Move compositional measurements by \code{shiftsize} towards the centre of the simplex.
}
\details{
Measurements on the boundary of the simplex can lead to degenerate log-likelihood values, which are avoided in score matching by the boundary weight function. However, \code{CppAD} still computes the log-likelihood and derivatives, and (to KLH's knowledge) is not able to evaluate the value of \eqn{0/0} or similar.
Good approximations of the score matching objective value can be obtained through Taylor approximations, which require a \emph{centre}.
The function here generates this approximation centre.
}
\examples{
m <- ppi_egmodel(10)
simplex_boundaryshift(m$sample, shiftsize = 1E-5)
}
