% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplex_bdryapprox.R
\name{simplex_isboundary}
\alias{simplex_isboundary}
\title{Determine whether points are on the simplex boundary}
\usage{
simplex_isboundary(Y, bdrythreshold = 1e-15)
}
\arguments{
\item{Y}{Measurement matrix.}

\item{bdrythreshold}{Measurements closer than \code{bdrythreshold} to the edge of the simplex will be considered boundary measurements.}
}
\value{
A vector of \code{TRUE} or \code{FALSE} values. Rows of \code{Y} corresponding to \code{TRUE} are on the boundary of the simplex.
}
\description{
Tests whether points are within \code{bdrythreshold} distance of the boundary, where distance is the size of minimum component.
}
\seealso{
Other simplex boundary helpers: 
\code{\link{simplex_boundaryshift}()}
}
\concept{simplex boundary helpers}
