% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smvalues_tape.R
\name{smvalues_tape}
\alias{smvalues_tape}
\alias{smvalues_tape_wsum}
\title{Compute score matching discrepancy value, gradient, and Hessian}
\usage{
smvalues_tape(smotape, xmat, pmat, xcentres = NA * xmat, approxorder = 10)

smvalues_tape_wsum(
  tape,
  xmat,
  pmat,
  w = NULL,
  xcentres = NA * xmat,
  approxorder = 10
)
}
\arguments{
\item{smotape}{A taped score matching discrepancy. Most easily created by \code{\link[=buildsmotape]{buildsmotape()}}.}

\item{xmat}{A matrix of (multivariate) independent variables. Each represents a single independent variable vector. Or a single vector of that is used for all rows of \code{pmat}.}

\item{pmat}{A matrix of dynamic parameters. Or a single vector of dynamic parameters to use for all rows of \code{xmat}.}

\item{xcentres}{A matrix of approximation for Taylor approximation centres for \code{xmat}. Use values of \code{NA} for rows that do not require Taylor approximation.}

\item{approxorder}{Order of Taylor approximation}

\item{tape}{A \code{CppAD} \code{tape}.}

\item{w}{Weights to apply to each row of \code{xmat} for computing the weighted sum. If \code{NULL} then each row is given a weight of \code{1}.}
}
\value{
A list of
\itemize{
\item \code{obj} the score matching discrepancy values
\item \code{grad} the gradient of the score matching discrepancy
\item \code{hess} the Hessian of the score matching discrepancy
}
}
\description{
Computes a range of relevant information for investigating score matching estimators.
}
\details{
The score matching discrepancy values are differ from the Hyv\"arinen Divergence by a constant, see ... .
The gradient and Hessian are returned as arrays of row-vectors with each row corresponding to a row in \code{xmat} and \code{pmat}.
}
\examples{
m <- ppi_egmodel(100)
tapes <- buildsmotape("sim", "sqrt", "sph", "ppi",
              ytape = rep(1/m$p, m$p),
              usertheta = ppi_paramvec(beta = m$beta),
              bdryw = "minsq", acut = 0.01)
smvalues_tape(tapes$smotape, xmat = m$sample, pmat = m$theta[1:5])
}
\seealso{
Other tape evaluators: 
\code{\link{evaltape_internal}},
\code{\link{evaltape}()},
\code{\link{quadratictape_parts}()},
\code{\link{testquadratictape}()}
}
\concept{tape evaluators}
