% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stiefel_utils.R
\name{Stiefel_embedding}
\alias{Stiefel_embedding}
\alias{vec}
\alias{invvec}
\alias{is_orthonormal}
\alias{in_Stiefel}
\title{Stiefel Manifold Embedding Tools}
\usage{
vec(A)

invvec(v, m)

is_orthonormal(A, tol = 1e-08)

in_Stiefel(v, m, tol = 1e-08)
}
\description{
An m x p matrix \code{A}, p <= m, is an element of the m,p Stiefel manifold if and only \code{t(A) \%*\% A == diag(p)}.
A Stiefel manifold using the matrix Frobenius norm as a metric can be embedded in an ambient mp Euclidean space  \insertCite{@Section 2.2, @edelman1998ge}{scorematchingad}
where each matrix \code{A} in the Stiefel manifold is represented by a mp-length vector \code{vec(A)} in Euclidean space.
}
\section{Functions}{
\itemize{
\item \code{vec()}: Stack columns of a matrix \code{A} on top of each other to obtain a vector of all elements of \code{A}. This wraps \code{as.vector()}, which does exactly the desired process.

\item \code{invvec()}: Break up a vector into columns of length \code{m}, and return as a matrix.

\item \code{is_orthonormal()}: Check is matrix is an element of Stiefel manifold.

\item \code{in_Stiefel()}: Check a vector represents an element of Stiefel manifold.

}}
