% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{swapDynamic}
\alias{swapDynamic}
\title{Switch Dynamic and Independent Values of a Tape}
\usage{
swapDynamic(pfun, newvalue, newdynparam)
}
\arguments{
\item{pfun}{An Rcpp::XPtr to an ADFun object (i.e. a tape of a function)}

\item{newvalue}{The independent value (in the sense after the switch has occurred) at which to tape the ADFun}

\item{newdynparam}{The value of the dynamic parameters (after the switch) at which to tape the ADFun}
}
\value{
A pointer to an ADFun
}
\description{
Convert an ADFun so that the independent values become dynamic parameters
and the dynamic parameters become independent values
}
