% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tape_smi.R
\name{tape_smi}
\alias{tape_smi}
\title{Build a Tape of the Score Matching Integrand}
\usage{
tape_smi(
  manifold,
  uld,
  transform = "none",
  xtape = uld$xtape,
  fixedparams = NA_real_ * uld$dyntape,
  bdryw = "ones",
  dynparam_filler = NULL,
  verbose = FALSE
)
}
\arguments{
\item{manifold}{A manifold object or name of a manifold used for estimation (which changes the score matching integrand)}

\item{uld}{Tape of an unnormalised log-density from \code{\link[=tape_uld_inbuilt]{tape_uld_inbuilt()}} or \code{\link[=tape_uld]{tape_uld()}}. For convenience, if \code{uld} is a string, it is passed to \code{\link[=tape_uld_inbuilt]{tape_uld_inbuilt()}}.}

\item{transform}{If the \code{uld} is defined on a different embedding to \code{manifold}, then the name of the transformation that maps to the embedding of \code{manifold}.}

\item{xtape}{Vector to use for taping the \code{uld}. If there is a \code{transform}, then the transform is also applied to \code{xtape}. If omitted, the value of \code{uld$xtape} is used.}

\item{fixedparams}{A vector of fixed parameters for (re)taping \code{uld}. \code{NA} elements will remain \emph{dynamic parameters}. Other elements will be fixed at the provided value. The same behaviour can also be obtained with \code{\link[=fixdynamic]{fixdynamic()}} and \code{\link[=fixindependent]{fixindependent()}}. If not specified, no parameters of \code{uld} are fixed. For getting the correct dimensions, \code{fixedparams} is required if \code{uld} is a string, however there is no checking that \code{fixedparams} is the correct length for \code{uld} and \code{xtape}.}

\item{bdryw}{The tape of a boundary weight function from \code{\link[=tape_bdryw]{tape_bdryw()}} or \code{\link[=tape_bdryw_inbuilt]{tape_bdryw_inbuilt()}}. For convenience use \code{bdryw="ones"} for manifolds without boundary. If \code{bdryw="prodsq"} or \code{bdryw="minsq"} then boundary weight function for the simplex and positive orthant of the sphere are created with default threshold - see \code{\link[=tape_bdrw_inbuilt]{tape_bdrw_inbuilt()}} for more information.}

\item{dynparam_filler}{Optional. Defaults to values of \code{uld$dyntape} unless \code{uld} is a string. A function that generates values to be used for taping the dynamic parameters remaining after \code{fixedparams} is applied. Must take a single argument, \code{n} the number for values to generate. When \code{uld} is a string, then defaults to \code{function(n){seq(length.out = n)}}.}

\item{verbose}{If \code{TRUE} more details are printed when taping. These details are for debugging and will likely be comprehensible only to users familiar with the source code of this package.}
}
\value{
A list of:
\itemize{
\item an \code{\link{Rcpp_ADFun}} object containing a tape of the unnormalised log-density using the metric of the "\code{end}" manifold (that is the independent variable is on the \code{end} manifold).
\item an \code{\link{Rcpp_ADFun}} object containing a tape of the score matching integrand with the non-fixed parameters of the model as the independent variable, and the measurements on the \code{end} manifold as the dynamic parameter.
\item some information about the tapes
}
}
\description{
For a parametric model family, the function \code{tape_smi()} generates \code{CppAD} tapes for the unnormalised log-density of the model family and of the score matching integrand \eqn{A(z) + B(z) + C(z)} (defined in \code{\link{scorematchingtheory}}).
Three steps are performed by \code{tape_smi()}: first an object that specifies the manifold and any transformation from another manifold is created; then a tape of the unnormalised log-density is created; finally a tape of \eqn{A(z) + B(z) + C(z)} is created.
}
\details{
If there is a \code{transform}, then \code{tape_smi()} first retapes via \code{\link[=reembed]{reembed()}} the \code{uld} for an independent variable \eqn{z} on \code{manifold} using the Jacobian provided in \code{transform} to account for the change in embedding (including the change of the uniform density/measure).
This tape has an independent variable \eqn{z} on \code{manifold}.
If supplied, \code{fixedparams} is applied using \code{\link[=fixdynamic]{fixdynamic()}}.
This tape is returned as \code{uld_reembed}.

This tape is then used to generate a tape for the score matching integrand where the parameters to estimate become the independent variable.
The dynamic parameters of the tape correspond to the original manifold (domain) of \code{uld}, which reflects the fact that the score matching integrand is altered by \code{transform}, not the model or data (the data is not transformed and the model is unchanged).
This tape is returned as \code{smi}.
}
\examples{
p <- 3
tapes <- tape_smi(manifold = "sph", 
                  uld = "vMF",
                  xtape = rep(1/sqrt(p), p),
                  fixedparams = rep(NA, p),
                  )
evaltape(tapes$uld_reembed, rep(1/sqrt(p), p), runif(n = p))
evaltape(tapes$smi, runif(n = p), rep(1/sqrt(p), p))

u <- rep(1/3, 3)
tapes <- tape_smi(manifold = "sph",
              uld = "ppi",
              transform = "sqrt"
              xtape = u,
              fixedparams = ppi_paramvec(p = 3), #no fixed params
              bdryw = "minsq" #uses default threshold for minsq
              )
evaltape(tapes$uld_reembed, sqrt(u), rppi_egmodel(1)$theta)
evaltape(tapes$smi, rppi_egmodel(1)$theta, u)
}
\references{
\insertAllCited{}
}
\seealso{
Other tape builders: 
\code{\link{avgrange}()},
\code{\link{fixdynamic}()},
\code{\link{fixindependent}()},
\code{\link{keeprange}()},
\code{\link{tape_Hessian}()},
\code{\link{tape_Jacobian}()},
\code{\link{tape_bdryw}()},
\code{\link{tape_gradoffset}()},
\code{\link{tape_logJacdet}()},
\code{\link{tape_smd}()},
\code{\link{tape_swap}()},
\code{\link{tape_uld}()}

Other generic score matching tools: 
\code{\link{Windham}()},
\code{\link{cppad_closed}()},
\code{\link{cppad_search}()},
\code{\link{tape_smd}()}
}
\concept{generic score matching tools}
\concept{tape builders}
