% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tape_swap}
\alias{tape_swap}
\title{Switch Dynamic and Independent Values of a Tape}
\usage{
tape_swap(pfun)
}
\arguments{
\item{pfun}{An \code{Rcpp_ADFun} object.}
}
\value{
An \code{Rcpp_ADFun} object.
}
\description{
Convert an \code{Rcpp_ADFun} object so that the independent values become dynamic parameters
and the dynamic parameters become independent values
}
\details{
For creating this tape, the values of \code{pfun$xtape} and \code{pfun$dyntape} are used.
}
\seealso{
Other tape builders: 
\code{\link{tape_Hessian}()},
\code{\link{tape_Jacobian}()},
\code{\link{tape_gradoffset}()},
\code{\link{tape_logJacdet}()},
\code{\link{tape_smd}()},
\code{\link{tape_uld}()}
}
\concept{tape builders}
