% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tapeJacobian}
\alias{tapeJacobian}
\title{Tape the Jacobian of CppAD Tape}
\usage{
tapeJacobian(pfun)
}
\arguments{
\item{pfun}{An \code{Rcpp_ADFun} object.}

\item{x}{A vector in the domain of the taped function.}

\item{dynparam}{a vector of the dynamic parameters.}
}
\value{
A \code{Rcpp::XPtr} to a \code{Rcpp_ADFun} object.
}
\description{
Creates a tape of the Jacobian of function taped by \code{CppAD}.
When the function returns a real value (as is the case for densities and the score matching objective) the Jacobian is equivalent to the gradient.
The \code{x} vector is used as the value to conduct the taping.
}
\details{
When the returned tape is evaluated (via say \verb{$eval()}, the resultant vector contains the Jacobian in long format (see \url{https://cppad.readthedocs.io/latest/Jacobian.html}).
Suppose the function represented by \code{pfun} maps from \eqn{n}-dimensional space to \eqn{m}-dimensional space, then
the first \eqn{n} elements of vector is the gradient of the first component of function output.
The next \eqn{n} elements of the vector is the gradient of the second component of the function output.
The Jacobian as a matrix, could then be obtained by \code{\link[=as.matrix]{as.matrix()}} with \code{byrow = TRUE} and \code{ncol = n}.
}
\seealso{
Other tape builders: 
\code{\link{buildsmdtape}()},
\code{\link{tapeGradOffset}()},
\code{\link{tapeHessian}()},
\code{\link{tapeLogJacDet}()},
\code{\link{tapeSwap}()}
}
\concept{tape builders}
