% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tapeLogJacDet}
\alias{tapeLogJacDet}
\title{Tape the log of Jacobian determinant of a CppAD Tape}
\usage{
tapeLogJacDet(pfun)
}
\arguments{
\item{pfun}{An \code{Rcpp_ADFun} object.}

\item{x}{A vector in the domain of the taped function.}

\item{dynparam}{a vector of the dynamic parameters (vector of size \code{0} if ther are no dynamic parameters).}
}
\value{
An \code{Rcpp_ADFun} object.
}
\description{
Creates a tape of the log of the Jacobian determinant of a function taped by CppAD.
The \code{x} vector is used as the value to conduct the taping.
}
\seealso{
Other tape builders: 
\code{\link{buildsmdtape}()},
\code{\link{tapeGradOffset}()},
\code{\link{tapeHessian}()},
\code{\link{tapeJacobian}()},
\code{\link{tapeSwap}()}
}
\concept{tape builders}
