% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tapesmo.R
\name{tapesmo}
\alias{tapesmo}
\title{Tape a score matching objective function}
\usage{
tapesmo(lltape, tran, man, divweight, acut = 1, verbose = FALSE)
}
\arguments{
\item{lltape}{Tape of the log-likelihood function from the manifold represented by \code{pmanifoldtransform}. Construct \code{lltape} using \code{\link[=tapell]{tapell()}}.}

\item{tran}{A transform object (of type \code{Rcpp_transform_ad}), typically created by \code{\link[=manifoldtransform]{manifoldtransform()}}.}

\item{man}{A transform object (of type \code{Rcpp_man_ad}) matching \code{tran}, and typically created by \code{\link[=manifoldtransform]{manifoldtransform()}}.}

\item{divweight}{The name of the divergence weight function ("ones" for manifolds without boundary). For the simplex and positive orthant of the sphere, "prodsq" and "minsq" are possible.}

\item{acut}{The threshold \eqn{a_c} in the divergence weight function \code{divweight}. Ignored for \code{divweight = "ones"}.}

\item{verbose}{If \code{TRUE}, some information about the tape is printed.}
}
\description{
Generates a \code{CppAD} tape of the empirical score matching objective function for a single measurement (\eqn{\tilde\psi} in Score Matching vignette). Requires a tape of the log-likelihood function and the corresponding manifold with the transformation.
}
\examples{
sqrtman <- manifoldtransform("sphere")
ppitape <- tapell(llname = "ppi",
                  ytape = c(0.2, 0.3, 0.5),
                  usertheta = ppi_paramvec(p = 3), 
                  pmanifoldtransform = sqrtman)
ppismotape <- tapesmo(lltape = ppitape,
                      pmanifoldtransform = sqrtman,
                      divweight = "minsq",
                      acut = 0.1,
                      verbose = FALSE)
pForward0(ppismotape, 
  ppi_paramvec(p = 3, AL=0, bL=0, beta=c(-0.1,-0.1,0.5)),
  rep(1/3, 3)) 
}
