% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tapeSwap}
\alias{tapeSwap}
\title{Switch Dynamic and Independent Values of a Tape}
\usage{
tapeSwap(pfun)
}
\arguments{
\item{pfun}{An \code{Rcpp_ADFun} object.}

\item{newvalue}{The independent value (in the sense after the switch has occurred) at which to tape the \code{Rcpp_ADFun}}

\item{newdynparam}{The value of the dynamic parameters (after the switch) at which to tape the \code{Rcpp_ADFun}}
}
\value{
An \code{Rcpp_ADFun} object.
}
\description{
Convert an \code{Rcpp_ADFun} object so that the independent values become dynamic parameters
and the dynamic parameters become independent values
}
\seealso{
Other tape builders: 
\code{\link{buildsmdtape}()},
\code{\link{tapeGradOffset}()},
\code{\link{tapeHessian}()},
\code{\link{tapeJacobian}()},
\code{\link{tapeLogJacDet}()}
}
\concept{tape builders}
