% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testquadratictape.R
\name{testquadratictape}
\alias{testquadratictape}
\title{Test Whether a CppAD Tape is a Quadratic Function}
\usage{
testquadratictape(tape, xmat = NULL, dynparammat = NULL, verbose = FALSE)
}
\arguments{
\item{tape}{An \code{Rcpp::XPtr} to a CppAD tape.}

\item{xmat}{If passed, the third-order derivatives at values of the rows of \code{xmat} are tested.}

\item{dynparammat}{The dynamic parameters for the tape. If passed, the rows of \code{dynparammat} are passed to the tape as \code{dynparam}.}

\item{verbose}{If TRUE information about the failed tests is passed.}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Uses \code{\link[=pParameter]{pParameter()}} and derivatives \code{\link[=pTapeJacobian]{pTapeJacobian()}}, \code{\link[=pJacobian]{pJacobian()}} to test whether
the tape is quadratic.
}
\details{
Uses the \code{xtape} and \code{dyntape} attributes of \code{tape} to create new tapes.
A tape of Hessian is obtained by applying \code{\link[=pTapeJacobian]{pTapeJacobian()}} twice. Using \code{\link[=pTapeHessian]{pTapeHessian()}} directly did not show constant parameters via \code{\link[=pParameter]{pParameter()}} in tests.

Two tests are conducted on the tape of the Hessian.
For a function of quadratic form, \code{\link[=pParameter]{pParameter()}} should return a vector of \code{TRUE} values.
The other test evaluates the Jacobian at each row of \code{xmat} and \code{dynparammat} (if \code{xmat} and \code{dynparammat} are non-NULL). The result should be zero when \code{tape} represents a function of quadratic form.
If the results of the tests differ the returned value is \code{FALSE} and a message is printed indicating which test failed.
}
\examples{
 sqrtman <- manifoldtransform("sphere")
 ppitape <- tapell(llname = "ppi",
                   xtape = c(0.2, 0.3, 0.5),
                   usertheta = ppi_paramvec(p = 3), 
                   pmanifoldtransform = sqrtman)
 ppismotape <- tapesmo(lltape = ppitape,
                       pmanifoldtransform = sqrtman,
                       divweight = "minsq",
                       acut = 0.1,
                       verbose = FALSE)

 testquadratictape(ppismotape)
}
