% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vMF.R
\name{vMF}
\alias{vMF}
\title{Score matching estimator for the von-Mises Fisher distribution}
\usage{
vMF(
  Y,
  paramvec = NULL,
  method = "smfull",
  control = default_Rcgmin(),
  w = rep(1, nrow(Y))
)
}
\arguments{
\item{Y}{A matrix of multivariate observations in Cartesian coordinates. Each row is a measurement.}

\item{paramvec}{\code{smfull} method only: Optional. A vector of same length as the dimension, representing the elements of the vector that is the element-wise (Hadamard) product \eqn{\kappa \mu}.}

\item{method}{Either "Mardia" for the hybrid score matching estimator from \insertCite{@mardia2016sc}{scorecompdir}.
or "smfull" for the full score matching estimator.}

\item{control}{Control parameters passed to \code{\link[Rcgmin:Rcgmin]{Rcgmin::Rcgmin()}}.}

\item{w}{An optional vector of weights for each measurement in \code{Y}}
}
\value{
A list of \code{est}, \code{SE} and \code{info}.
\itemize{
\item \code{est} contains the estimates in vector form, \code{paramvec}, and with user friendly names \code{k} and \code{m}.
\item \code{SE} contains estimates of the standard errors if computed.
\item \code{info} contains a variety of information about the model fitting procedure and results.
}
}
\description{
We recommend using \code{\link[movMF:movMF]{movMF::movMF()}} over the function here, \code{vMF()}, for estimating the von Mises Fisher distribution.

\insertCite{mardia2016sc;textual}{scorecompdir} suggest score matching for the von Mises Fisher distribution to avoid the difficult-to-compute normalising constant of the von Mises Fisher distributions.
\insertCite{mardia2016sc;textual}{scorecompdir} suggested both full score matching and a hybrid estimator, where by the mean direction is estimated via the maximum-likelihood approach and the concentration is estimated via score matching.
However, \insertCite{mardia2016sc;textual}{scorecompdir} found, that these score matching estimators were not as a efficient as the maximum likelihood estimators (e.g. \code{\link[movMF:movMF]{movMF::movMF()}} ) on the circle.
Some of our own experiments suggest this is also the case for higher dimensions, despite the more complex approximations required for the maximum likelihood estimators.
}
\details{
The method "smfull" uses score matching to estimate the vector \eqn{\kappa \mu}.
The method "Mardia" uses \code{\link[=vMF_stdY]{vMF_stdY()}} and \code{\link[=vMF_kappa]{vMF_kappa()}} to estimate \eqn{\kappa} and \eqn{\mu} seperately.
}
\section{von Mises Fisher Model}{

The von Mises Fisher density is proportional to
\deqn{\exp(\kappa \mu^T z),}
where \eqn{z} is on a unit sphere,
\eqn{\kappa} is termed the \emph{concentration},
and \eqn{\mu} is the \emph{mean direction vector}.
The effect of the \eqn{\mu} and \eqn{\kappa} can be decoupled in a sense \insertCite{@p169, @mardia2000di}{scorecompdir}, allowing for estimating \eqn{\mu} and \eqn{\kappa} separately.
}

\examples{
set.seed(12342)
Y <- movMF::rmovMF(1000, 100 * c(1, 1) / sqrt(2))
movMF::movMF(Y, 1) #maximum likelihood estimate
vMF(Y, method = "smfull")
vMF(Y, method = "Mardia")
}
\references{
\insertAllCited{}
}
\seealso{
Other directional model estimators: 
\code{\link{Bingham}()},
\code{\link{FB}()},
\code{\link{vMF_kappa}()},
\code{\link{vMF_m}()},
\code{\link{vMF_robust}()}
}
\concept{directional model estimators}
