% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vMF_kappa.R
\name{vMF_kappa}
\alias{vMF_kappa}
\title{Estimate the concentration for a von Mises Fisher distribution}
\usage{
vMF_kappa(Y, w = rep(1, nrow(Y)), control = default_Rcgmin())
}
\arguments{
\item{Y}{A data matrix, each row is an observation.}

\item{w}{Weights corresponding to each row of \code{Y}.}

\item{control}{A list of control arguments passed to \code{\link[Rcgmin:Rcgmin]{Rcgmin::Rcgmin()}}.}
}
\value{
A list of \code{est}, \code{SE} and \code{info}.
\code{est} contains the estimate of the concentration in the slot \code{k} for easy use and \code{paramvec} for compatibility with other functions in this package.
\code{SE} contains estimates of the standard errors if computed by the estimating method.
\code{info} contains a variety of information about the model fitting procedure.
}
\description{
Using score matching, estimates the concentration \eqn{\kappa} from a sample with mean direction of \code{c(1, 0, 0, ..., 0)}.
Results by \insertCite{mardia2016sc;textual}{scorecompdir} and some experiments of our own suggest that good implementations of the maximum likelihood estimator (e.g. \code{\link[movMF:movMF]{movMF::movMF()}} ) will out perform \code{vMF_kappa()}.

Often a sample with mean direction of \code{c(1, 0, 0, ...., 0)} is created by estimating the mean direction and rotating the data such that the mean direction equals \code{c(1, 0, 0, ...)}.
Performing this mean direction estimate, then rotation, then estimating concentration with score matching correponds to the hybrid estimator by \insertCite{mardia2016sc;textual}{scorecompdir}.
}
\details{
The function \code{vMF_kappa()} estimates \emph{only} the concentration \eqn{\kappa}, and assumes that \eqn{\mu} is \eqn{(1, 0, 0, ..., 0)}.
}
\section{von Mises Fisher Model}{

The von Mises Fisher density is proportional to
\deqn{\exp(\kappa \mu^T z),}
where \eqn{z} is on a unit sphere,
\eqn{\kappa} is termed the \emph{concentration},
and \eqn{\mu} is the \emph{mean direction vector}.
The effect of the \eqn{\mu} and \eqn{\kappa} can be decoupled in a sense \insertCite{@p169, @mardia2000di}{scorecompdir}, allowing for estimating \eqn{\mu} and \eqn{\kappa} separately.
}

\references{
\insertAllCited()
}
\seealso{
Other Mardia hybrid estimators: 
\code{\link{Bingham}()},
\code{\link{vMF}()}
}
\concept{Mardia hybrid estimators}
