% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vMF_m.R
\name{vMF_m}
\alias{vMF_m}
\alias{vMF_stdY}
\title{Mean direction and rotation for von Mises Fisher}
\usage{
vMF_m(Y, w = NULL)

vMF_stdY(Y, m = NULL, w = NULL)
}
\arguments{
\item{Y}{A matrix of observations in Cartesian coordinates. Each row is a single observation.}

\item{w}{Weights associated with each observation in \code{Y}.}

\item{m}{Mean direction. If omitted then it will be computed using \code{vMF_m()}.}
}
\description{
Computes the mean direction of a sample, which is also the maximum likelihood estimate for the mean direction parameter \code{m} in the von Mises Fisher distribution.
}
\details{
\code{vMF_m()} Rotates a set of observations on a sphere to have mean direction of \eqn{(1, 0, 0, ..., 0)}.
}
\examples{
Y <- simdd::rFisherBingham(nsim = 10, mu = c(1,2,3))
}
\seealso{
Other directional model estimators: 
\code{\link{Bingham}()},
\code{\link{FB}()},
\code{\link{vMF_kappa}()},
\code{\link{vMF_robust}()},
\code{\link{vMF}()}
}
\concept{directional model estimators}
