% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vMF_robust.R
\name{vMF_robust}
\alias{vMF_robust}
\title{Robust fitting of von Mises Fisher}
\usage{
vMF_robust(Y, cW, ...)
}
\arguments{
\item{Y}{A matrix of observations in Cartesian coordinates.}

\item{cW}{Tuning constants for each parameter in the vMF parameter vector. If a single number then the constant is the same for each element of the parameter vector.}

\item{...}{Passed to \code{\link[=Windham]{Windham()}} and then passed onto \code{\link[=vMF]{vMF()}}.}
}
\description{
Robust estimation for von Mises Fisher distribution using \code{\link[=Windham]{Windham()}}.
}
\seealso{
Other directional model estimators: 
\code{\link{Bingham}()},
\code{\link{FB}()},
\code{\link{vMF}()}

Other Windham robustness functions: 
\code{\link{Windham}()},
\code{\link{ppi_robust}()}
}
\concept{Windham robustness functions}
\concept{directional model estimators}
