% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Windham_weights.R
\name{Windham_weights}
\alias{Windham_weights}
\title{Windham weights for a given parameter vector}
\usage{
Windham_weights(ldenfun, Y, theta, cW)
}
\arguments{
\item{ldenfun}{A (possibly improper) log density function taking two arguments, \code{Y} and \code{theta}.}

\item{Y}{A matrix of measurements. Each row a measurement. Passed to \code{ldenfun}.}

\item{theta}{Parameters for the model. The element-wise multiplication \code{cW * theta} is passed to \code{ldenfun}.}

\item{cW}{The robustness tuning constants. A vector of the same length as \code{theta}. Easily created for the PPI model using \code{\link[=ppi_cW]{ppi_cW()}} and \code{\link[=ppi_cW_auto]{ppi_cW_auto()}}.}
}
\value{
A vector of weights corresponding to the rows of \code{Y}.
The weights are normalised to sum to 1.
}
\description{
Evaluates \eqn{f(z, c\circ\theta)}, which is the density \eqn{f} at observation \eqn{z}, given a parameter set \eqn{\theta} and vector of tuning constants \eqn{c}. The multiplication \eqn{\circ} is element-wise.
These are the weights used by \code{\link[=Windham]{Windham()}}.
The density function is passed to \code{Windham_weights()} in log form.
}
\seealso{
Other Windham functions: 
\code{\link{Windham}()},
\code{\link{ppi_robust}()},
\code{\link{vMF_robust}()}
}
\concept{Windham functions}
