% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifoldtransform.R
\name{manifoldtransform}
\alias{manifoldtransform}
\title{Build an object specifying the manifold-transform pair.}
\usage{
manifoldtransform(name)
}
\arguments{
\item{name}{Name of the manifold-transform pair. See details.}
}
\value{
An \code{Rcpp::XPtr} object with attribute \code{name}.
}
\description{
Generate an object used to specify a manifold and transformation to the manifold. This specific to each type of data.
}
\details{
Available pairs are
\itemize{
\item "sphere" for square-root transformation from the simplex to the positive orthant of the sphere
\item "simplex" for the simplex without any transformation.
\item "Ralr" for the additive log-ratio transformation from the simplex to Euclidean space, using the final component of vectors in the denominator of the ratio.
\item "Snative" for the sphere without any transformation
}
}
\examples{
manifoldtransform("Ralr")
}
