% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tape_uld.R
\name{tape_uld}
\alias{tape_uld}
\title{Generate a tape of custom unnormalised log-density}
\usage{
tape_uld(file = "", x, theta, Cppopt = NULL)
}
\arguments{
\item{file}{A character string giving the path name of a file containing the unnormalised log-density definition.}

\item{x}{Value of \code{x} for taping.}

\item{theta}{Value of the parameter vector for taping.}

\item{Cppopt}{List of named options passed to \code{Rcpp::sourceCpp()}}
}
\value{
A list of the tape and the function itself
}
\description{
Use Rcpp::sourceCpp to generate a tape of a function defined in C++. The result is NOT safe to save or pass to other CPUs in a parallel operation.
}
\examples{
tape_uld(system.file("demo_custom_uld.cpp", package = "scorematchingad"), rep(0.2, 5), rep(-0.1, 5))
}
