% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tapell.R
\name{tapell}
\alias{tapell}
\title{Tape a log-likelihood}
\usage{
tapell(
  llname,
  ytape,
  usertheta,
  tran,
  thetatape_creator = function(n) {
     seq(length.out = n)
 },
  verbose = FALSE
)
}
\arguments{
\item{llname}{The name of the log-likelihood function to tape}

\item{ytape}{An example measurement value to use for creating the tape. In the natural manifold of the log-likelihood function. \code{ytape} will be converted to the manifold according to the \code{toM()} method for \code{tran} before taping.
Please ensure that \code{ytape} is the interior of the manifold, and it is probably best if all components of \code{tran$toM(ytape)} are non-zero.}

\item{usertheta}{A vector of parameter elements for the likelihood function. \code{NA} elements will become \emph{dynamic parameters}. Other elements will be fixed at the provided value. The length of \code{usertheta} must be the correct length for the log-likelihood, no checking is conducted.}

\item{tran}{A transform object (of type \code{Rcpp_transform_ad}), typically created by \code{\link[=manifoldtransform]{manifoldtransform()}}.}

\item{thetatape_creator}{A function that accepts an integer \code{n}, and returns a vector of \code{n} length. The function is used to fill in the \code{NA} elements of \code{usertheta} when building the tapes. Please ensure that the values filled by \code{thetatape_creator} lead to plausible parameter vectors for the chosen log-likelihood.}

\item{verbose}{If \code{TRUE}, some information about the tape is printed.}
}
\value{
An \code{Rcpp::XPtr} object that points to a \code{CppAD::ADFun} object.
The object will also have attributes:
\itemize{
\item \code{fname} A name of the taped function, derived from \code{llname}
\item \code{ytape} The value of \code{ytape}
\item \code{dyntape} The value of non-fixed elements of \code{usertheta} used for taping.
\item \code{usertheta} The vector \code{usertheta}.
}
}
\description{
Creates a \code{CppAD} tape of the log-likelihood as a function of values on the provided manifold. The log-likelihood accounts for the change of measure between the natural manifold of the log-likelihood and the manifold with transformation provided by \code{manifoldname}.
}
\section{Warning}{
 There is limited checking of the inputs.
}

\examples{
simpwsqrt <- manifoldtransform("sphere")
ppitape <- tapell(llname = "ppi",
                  ytape = c(0.2, 0.3, 0.5),
                  usertheta = ppi_paramvec(p = 3), 
                  pmanifoldtransform = simpwsqrt)
pForward0(ppitape, 
  sqrt(rep(1/3, 3)), 
  ppi_paramvec(p = 3, AL=0, bL=0, beta=c(0,0,0.5)))
pJacobian(ppitape, 
  sqrt(rep(1/3, 3)), 
  ppi_paramvec(p = 3, AL=0, bL=0, beta=c(0,0,0.5)))
pHessian(ppitape, 
  sqrt(rep(1/3, 3)), 
  ppi_paramvec(p = 3, AL=0, bL=0, beta=c(0,0,0.5)))
}
