% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tapeSwap}
\alias{tapeSwap}
\title{Switch Dynamic and Independent Values of a Tape}
\usage{
tapeSwap(pfun)
}
\arguments{
\item{pfun}{An \code{Rcpp_ADFun} object.}
}
\value{
An \code{Rcpp_ADFun} object.
}
\description{
Convert an \code{Rcpp_ADFun} object so that the independent values become dynamic parameters
and the dynamic parameters become independent values
}
\details{
For creating this tape, the values of \code{pfun$xtape} and \code{pfun$dyntape} are used.
}
\seealso{
Other tape builders: 
\code{\link{buildsmdtape}()},
\code{\link{tapeGradOffset}()},
\code{\link{tapeHessian}()},
\code{\link{tapeJacobian}()},
\code{\link{tapeLogJacDet}()},
\code{\link{tape_uld_inbuilt}()}
}
\concept{tape builders}
