% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vMF_robust.R
\name{vMF_robust}
\alias{vMF_robust}
\alias{vMF_kappa_robust}
\title{Robust fitting of von Mises Fisher}
\usage{
vMF_robust(Y, cW, ...)

vMF_kappa_robust(Y, cW, ...)
}
\arguments{
\item{Y}{A matrix of observations in Cartesian coordinates.}

\item{cW}{Tuning constants for each parameter in the vMF parameter vector. If a single number then the constant is the same for each element of the parameter vector.}

\item{...}{Passed to \code{\link[=Windham]{Windham()}} and then passed onto either \code{\link[=vMF]{vMF()}} or \code{\link[=vMF_kappa]{vMF_kappa()}}.}
}
\description{
Robust fitting for von Mises Fisher distribution using \code{\link[=Windham]{Windham()}},
for the full parameter vector \eqn{\kappa\mu} with \code{vMF_robust()} or just the concentration with \code{vMF_kappa_robust()} assuming a mean direction of \code{c(1, 0, 0, ....)}.
}
\seealso{
Other directional model estimators: 
\code{\link{Bingham}()},
\code{\link{FB}()},
\code{\link{vMF_kappa}()},
\code{\link{vMF_m}()},
\code{\link{vMF}()}

Other Windham robustness functions: 
\code{\link{Windham}()},
\code{\link{ppi_robust_alrgengamma}()},
\code{\link{ppi_robust}()}
}
\concept{Windham robustness functions}
\concept{directional model estimators}
